separated guards into different files
