+++ client/src/App.js
@@ -5,7 +5,8 @@ import {Routes, Route, Outlet} from 'react-router-dom';
 import CreateAccount from './CreateAccount';
 import Worlds from './Worlds';
 import Login from './Login';
-import { AdminRoute, ProtectedRoute} from "./route-guards";
+import { AdminRoute } from './route-guards/admin-guard';
+import { ProtectedRoute } from './route-guards/login-guard';
 import HomeLoggedOut from './Home-LoggedOut';
 import Navbar from './Navbar';
 import Profile from './Profile';
