+++ client/src/route-guards/login-guard.js
@@ -1,0 +1,11 @@
+import { checkAuth } from "../services/auth-service";
+import {Navigate} from "react-router-dom";
+
+export const ProtectedRoute = ({ children }) => {
+    const user = checkAuth()
+    if (!user) {
+      // user is not authenticated
+      return <Navigate to="/login" />;
+    }
+    return children;
+};
